refactor: :art: rename component folder names
