+++ src/app/app-routing.module.ts
@@ -1,9 +1,9 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { HomeComponent } from './components/home/home.component';
-import { LoginPageComponent } from './components/loginPage/loginPage.component';
-import { RegisterPageComponent } from './components/registerPage/registerPage.component';
-import { ReportComponent } from './components/report/report.component';
+import { HomeComponent } from './components/home-page/home.component';
+import { LoginPageComponent } from './components/login-page/loginPage.component';
+import { RegisterPageComponent } from './components/register-page/registerPage.component';
+import { ReportComponent } from './components/report-page/report.component';
 
 const routes: Routes = [
     { path: '', component: HomeComponent },
