+++ src/app/components/home-page/home.component.css
@@ -6,6 +6,7 @@ h3 {
     margin-bottom: 0;
 }
 
-.PGA mat-card-title, .PGA p {
+.PGA mat-card-title,
+.PGA p {
     text-align: center;
 }
