--- src/app/components/report/report.component.css
@@ -1,85 +1,0 @@
-mat-card {
-    margin: 10px 10px;
-}
-
-h3 {
-    margin-bottom: 0;
-}
-
-mat-card.PGA  {
-    padding: 15px;
-}
-
-.PGA p {
-    text-align: center;
-}
-
-.PGA mat-card-title p {
-    color: black;
-    font-size: 26pt;
-}
-
-mat-card-title {
-    color: var(--secundary);
-}
-
-mat-form-field {
-    width: -webkit-fill-available;
-    margin: 0 10px;
-}
-
-.reportRow1 mat-card {
-    height: 110px;
-    vertical-align: middle;
-}
-
-.credits {
-    padding: 7px;
-    width: 40px;
-    height: 40px;
-    text-align: center;
-    font-size: 16pt;
-    vertical-align: bottom;
-    border: 2px solid darkgray;
-    border-radius: 50px;
-    font-weight: bold;
-    float: right;
-}
-
-.doubleSeparator {
-    margin: 0;
-}
-
-.doubleSeparator p{
-    display: inline-block;
-    margin-bottom: 0;
-}
-
-.grade {
-    float: right;
-    padding: 7px;
-    width: 60px;
-    height: 40px;
-    text-align: center;
-    font-size: 16pt;
-    vertical-align: bottom;
-    border: 2px solid darkgray;
-    border-radius: 10px;
-    margin-top: -15px;
-    font-weight: bold;
-}
-
-.subject mat-card-title, .subject mat-card-subtitle {
-    margin-left: -16px;
-}
-
-.sumCredits p {
-    vertical-align: middle;
-    font-weight: bold;
-    font-size: 18pt;
-}
-
-.sumCredits div {
-    display: flex;
-    align-items: center;
-}
